zurb-foundation base-theme moved out of overall area
This is because it cause issues when jumping versions. CIS core distro
currently uses Zurb 4 but with Zurb 5 now being stable, Collins pointed
out developing in Z5 for new things makes sense but old things might
get busted. This will have to be added per stack which is redundant and
annoying but not too bad. Bootstrap will get this treatment soon, as
will anything using a base-theme.