site upgrade routine abstracted from full system
This allows for git pull to affect this file if there are changes to it
without it needing to be run twice to actually enact changes. While
this is true of the system upgrade script as well, drush / drupal
implications of upgrades being applied are higher priority then the
structure of sending data to remote services, etc.