ELMSLN symlink resyc and add-stack drush calls added
These are for local development when your on the development team.
Right now, that’s just me so you ever get the sense that i’m trying to
eventually hook literally everything I do up to a robot to do it for
me. I mean honestly, is there a reason for me to not go and run the 5-6
symlink calls that some of this would requite? No, probably not, but
when you are trying to design sustainably that means you try and
automated yourself out of existence in all ways. Who knows, maybe
someday this gem of a commit message will also be auto produced by the
ELMSLN-AI agent that I write to add my snark and overly detailed commit
conversations automatically through analysis of past ones. It could
happen…