#942 ELMSLN_CDN module update
Added JSDelivr Support for the ELMSLN CDN module. Checks to see if library can be found by the `libraries_detect()` function. Then sees if it can match a version found on the CDN. If both match. Then it finds all the files that are on the CDN and sets the URL to be the 'data' property of the javascript array.