Added local Angular Libraries Support #405 & #978
Added local support for Angular libraries. We will be using version 1.5.7 for now to start. This is now a library that is detectable by the Libraries API module and will eventually be able to be used with our CDN module (#991). We will need to run the `drush elms-sync-libs` command in order to hook the library into all the systems. This has the full angular library - so if we ever want to cut down on size, we can figure out which files we need / want to load.