#589 Features Export
Did a re-export of the features for #589. Looks cleaner than the last one, but I’m not sure if the deletion of textbook_editor is a mistaken deletion? There is also reference to a mooc_helper views default function being nixed. If these are incorrect I’ll do a re-export and try to re-add them in.